Fix a bug with duplicate pathnames
When you define a pathname in the router, proxy requests duplicate this
pathname.

For example, with a router.json of:

{
  "router": {
    "routes": {
      "/elasticsearch": {
        "protocol": "http:",
        "hostname": "localhost",
        "port": 8081,
        "pathname": "/foo"
      }
    }
  }
}

a proxy request to https://guardian/elasticsearch/bar will get routed to
http://localhost:8081/foo/foo/bar.

The pathname is already being merged as part of Client.mergeOptions, so
there is no reason to do this as part of the Proxy request.